Create common component for address editor
